requirements.txt
django
PyJWT
asgiref
cachetools
certifi
chardet
djangorestframework
djangorestframework-simplejwt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
oauthlib
packaging
pip
protobuf
pyasn1
pyasn1-modules
pyparsing
python-dotenv
pytz
requests
requests-oauthlib
rsa
setuptools
six
sqlparse
uritemplate
urllib3
bs4
Pillow
psycopg2
gunicorn==20.1.0
whitenoise==5.2.0
